@@ -1,7 +1,9 @@
 import expect from 'expect'
-import React, { createClass, Children, PropTypes, Component } from 'react'
+import React, { Children, Component } from 'react'
+import createClass from 'create-react-class'
+import PropTypes from 'prop-types'
 import ReactDOM from 'react-dom'
-import TestUtils from 'react-addons-test-utils'
+import TestUtils from 'react-dom/test-utils'
 import { createStore } from 'redux'
 import { connect } from '../../src/index'
 
@@ -9,7 +11,7 @@ describe('React', () => {
   describe('connect', () => {
     class Passthrough extends Component {
       render() {
-        return <div {...this.props} />
+        return <div />
       }
     }
 
@@ -1381,7 +1383,7 @@ describe('React', () => {
       }
 
       ImpureComponent.contextTypes = {
-        statefulValue: React.PropTypes.number
+        statefulValue: PropTypes.number
       }
 
       const decorator = connect(state => state, null, null, { pure: false })
@@ -1405,7 +1407,7 @@ describe('React', () => {
       }
 
       StatefulWrapper.childContextTypes = {
-        statefulValue: React.PropTypes.number
+        statefulValue: PropTypes.number
       }
 
       const tree = TestUtils.renderIntoDocument(
@@ -1727,7 +1729,7 @@ describe('React', () => {
           updatedCount++
         }
         render() {
-          return <div {...this.props} />
+          return <div />
         }
       }
 
